% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_total_or_average_distances.R
\name{calculate_set_distances}
\alias{calculate_set_distances}
\title{Calculate average distance in a given set}
\usage{
calculate_set_distances(distance_matrix, set, unit_distances = FALSE,
  sum_distances = FALSE, min_distance = FALSE)
}
\arguments{
\item{distance_matrix}{A distance matrix (see create_distance_matrix()).}

\item{set}{A set of unit IDs}

\item{unit_distances}{If unit_distances = TRUE, calculates distances at unit level.}

\item{sum_distances}{If sum_distances = TRUE, sums instead of averaging.}
}
\description{
Calculate average distance in a given set
}

