% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_spherical_distance.R
\name{calculate_spherical_distance}
\alias{calculate_spherical_distance}
\title{Calculate the spherical distance between two coordinates}
\usage{
calculate_spherical_distance(long1, lat1, long2, lat2, km = NULL,
  miles = NULL, meters = NULL, feet = NULL, ...)
}
\arguments{
\item{long1}{Longitude of point 1}

\item{lat1}{Latitude of point 1}

\item{long2}{Longitude of point 2}

\item{lat2}{Latitude of point 2}

\item{km}{Express distance in kilometers}

\item{miles}{Express distance in miles}

\item{meters}{Express distance in meters}

\item{feet}{Express distance in feet}
}
\value{
The spherical distance between point 1 and point 2.
}
\description{
Calculate the spherical distance between two coordinates
}

