% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_total_or_average_distances.R
\name{calculate_total_or_average_distances}
\alias{calculate_total_or_average_distances}
\title{Calculate the total or average distance between all units for each set in a sample_selection object}
\usage{
calculate_total_or_average_distances(sample_selection_object, distance_matrix,
  total = TRUE)
}
\arguments{
\item{sample_selection_object}{A list of admissable samples, created by a create_set() function (create_set_spatial_buffer() etc.)}

\item{distance_matrix}{A distance matrix, see function create_distance_matrix()}

\item{total}{If total = TRUE, the total distance between all units is returned, if total=FALSE, the average distance between all units is returned}
}
\description{
Calculate the total or average distance between all units for each set in a sample_selection object
}

