% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_adjacency_matrix.R
\name{create_adjacency_matrix}
\alias{create_adjacency_matrix}
\title{Create a binary adjacency matrix from a distance matrix.}
\usage{
create_adjacency_matrix(distance_matrix, threshold, zero_diagonal = T, ...)
}
\arguments{
\item{distance_matrix}{A distance matrix (see create_distance_matrix()).}

\item{threshold}{The minimum distance one unit should be from another (expressed in same units as distance matrix)}

\item{zero_diagonal}{A logical argument indicating whether the diagonal of the adjacency matrix should be all 0's, such that a unit is not defined as adjacent to itself.}
}
\value{
An adjacency matrix with binary indicators for adjacency.
}
\description{
Create a binary adjacency matrix from a distance matrix.
}

