% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_distance_matrix.R
\name{create_distance_matrix}
\alias{create_distance_matrix}
\title{Calculate the distance of each unit to every other unit.}
\usage{
create_distance_matrix(latitudes, longitudes, ...)
}
\arguments{
\item{latitudes}{A vector of latitudes of length N}

\item{longitudes}{A vector of longitudes of length N}

\item{...}{Takes distances arguments. Set km, meters, miles or feet to true in order to change the measure of distance that is output by the function.}
}
\value{
An N x N distance matrix
}
\description{
Calculate the distance of each unit to every other unit.
}

