% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_total_or_average_distances.R
\name{dist_mat_subset}
\alias{dist_mat_subset}
\title{Subset a distance matrix by a sample_selection object}
\usage{
dist_mat_subset(distance_matrix, set)
}
\arguments{
\item{distance_matrix}{A distance matrix (see create_distance_matrix()).}

\item{sample_selection}{A sample_selection object (see find_sets()).}
}
\value{
A subsetted distance matrix.
}
\description{
Subset a distance matrix by a sample_selection object
}

