% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_set_spatial_buffer.R
\name{find_proximate_units}
\alias{find_proximate_units}
\title{For each unit, find the units that are within some distance}
\usage{
find_proximate_units(distance_matrix, threshold, ...)
}
\arguments{
\item{distance_matrix}{A distance matrix (see create_distance_matrix()).}

\item{threshold}{The minimum distance one unit should be from another (expressed in same units as distance matrix)}
}
\value{
A list of vectors indicating, for each unit, the units within the spatial threshold.
}
\description{
For each unit, find the units that are within some distance
}

