% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_sets.R
\name{find_sets_biggest}
\alias{find_sets_biggest}
\title{Find the biggest sets within a sample_selection object}
\usage{
find_sets_biggest(sample_selection_object)
}
\arguments{
\item{sample_selection_object}{A list of admissable samples, created by a create_set() function (create_set_spatial_buffer() etc.)}
}
\value{
A list of the biggest admissable sets.
}
\description{
Find the biggest sets within a sample_selection object
}

