% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_sets.R
\name{find_sets_total_distance}
\alias{find_sets_total_distance}
\title{Find the sets that are above a given percentile of the distribution of total distances between units within a sample_selection object
#'}
\usage{
find_sets_total_distance(sample_selection_object, distance_matrix, prob = 1)
}
\arguments{
\item{sample_selection_object}{A list of admissable samples, created by a create_set() function (create_set_spatial_buffer() etc.)}

\item{distance_matrix}{A distance matrix, see function create_distance_matrix()}

\item{prob}{A probability, the function finds all sets that fall above the corresponding quantile of the distribution of total distances between units among the sets within a sample_selection object}
}
\value{
A list of the  admissable sets that maximize the distance between the units in the set.
}
\description{
Find the sets that are above a given percentile of the distribution of total distances between units within a sample_selection object
#'
}

