% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_total_or_average_distances.R
\name{summarize_set_distances}
\alias{summarize_set_distances}
\title{Summarize the distances in a set of sets}
\usage{
summarize_set_distances(distance_matrix, sample_selection,
  unit_distances = FALSE, sum_distances = FALSE)
}
\arguments{
\item{distance_matrix}{A distance matrix (see create_distance_matrix()).}

\item{sample_selection}{A sample_selection object (see find_sets()).}

\item{unit_distances}{If unit_distances = TRUE, calculates distances at unit level.}

\item{sum_distances}{If sum_distances = TRUE, sums instead of averaging.}

\item{unit}{If unit == TRUE, calculates distances at unit level.}
}
\description{
Summarize the distances in a set of sets
}

